+++ src/main/java/cart/application/CartItemService.java
@@ -4,10 +4,10 @@ import cart.dao.CartItemDao;
 import cart.dao.ProductDao;
 import cart.domain.CartItem;
 import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
-import cart.dto.RemoveCartItemsRequest;
+import cart.ui.dto.request.CartItemQuantityUpdateRequest;
+import cart.ui.dto.request.CartItemRequest;
+import cart.ui.dto.request.RemoveCartItemsRequest;
+import cart.ui.dto.response.CartItemResponse;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
