+++ src/main/java/cart/application/OrderService.java
@@ -11,12 +11,12 @@ import cart.domain.Order;
 import cart.domain.Product;
 import cart.domain.Products;
 import cart.domain.vo.Amount;
-import cart.dto.CartItemRequest;
-import cart.dto.OrderListResponse;
-import cart.dto.OrderProductResponse;
-import cart.dto.OrderRequest;
-import cart.dto.OrderResponse;
 import cart.exception.BusinessException;
+import cart.ui.dto.request.CartItemRequest;
+import cart.ui.dto.request.OrderRequest;
+import cart.ui.dto.response.OrderListResponse;
+import cart.ui.dto.response.OrderProductResponse;
+import cart.ui.dto.response.OrderResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
