+++ src/main/java/cart/application/ProductService.java
@@ -3,8 +3,8 @@ package cart.application;
 import cart.dao.ProductDao;
 import cart.domain.Product;
 import cart.domain.vo.Amount;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.ui.dto.request.ProductRequest;
+import cart.ui.dto.response.ProductResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
