+++ src/main/java/cart/ui/CouponApiController.java
@@ -2,8 +2,8 @@ package cart.ui;
 
 import cart.application.CouponService;
 import cart.domain.Member;
-import cart.dto.CouponDiscountResponse;
-import cart.dto.CouponResponse;
+import cart.ui.dto.response.CouponDiscountResponse;
+import cart.ui.dto.response.CouponResponse;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
