+++ src/main/java/cart/ui/dto/request/RemoveCartItemsRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.ui.dto.request;
 
 import java.util.List;
