+++ src/main/java/cart/ui/dto/response/CouponDiscountResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.ui.dto.response;
 
 public class CouponDiscountResponse {
