+++ src/test/java/cart/application/CouponServiceTest.java
@@ -10,10 +10,10 @@ import cart.dao.CouponDao;
 import cart.domain.Coupon;
 import cart.domain.Member;
 import cart.domain.vo.Amount;
-import cart.dto.CouponDiscountResponse;
-import cart.dto.CouponResponse;
-import cart.dto.PossibleCouponResponse;
 import cart.exception.BusinessException;
+import cart.ui.dto.response.CouponDiscountResponse;
+import cart.ui.dto.response.CouponResponse;
+import cart.ui.dto.response.PossibleCouponResponse;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
