+++ src/test/java/cart/integration/CouponApiControllerTest.java
@@ -8,8 +8,8 @@ import cart.dao.MemberDao;
 import cart.domain.Coupon;
 import cart.domain.Member;
 import cart.domain.vo.Amount;
-import cart.dto.CouponDiscountResponse;
-import cart.dto.ProductRequest;
+import cart.ui.dto.request.ProductRequest;
+import cart.ui.dto.response.CouponDiscountResponse;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.BeforeEach;
