+++ src/test/java/cart/integration/OrderApiControllerTest.java
@@ -8,11 +8,11 @@ import cart.dao.MemberDao;
 import cart.domain.Coupon;
 import cart.domain.Member;
 import cart.domain.vo.Amount;
-import cart.dto.CartItemRequest;
-import cart.dto.OrderProductResponse;
-import cart.dto.OrderRequest;
-import cart.dto.OrderResponse;
-import cart.dto.ProductRequest;
+import cart.ui.dto.request.CartItemRequest;
+import cart.ui.dto.request.OrderRequest;
+import cart.ui.dto.request.ProductRequest;
+import cart.ui.dto.response.OrderProductResponse;
+import cart.ui.dto.response.OrderResponse;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
